+++ Changelog.md
@@ -125,6 +125,7 @@
 -   Improved the loading times of the "Ticks" tab by removing an unnecessary sort operation on the list of pending tick keys.
 -   The "Ticking Areas" tab now no longer unnecessarily loads the dynamic properties data.
 -   The "Structures" tab now no longer unnecessarily loads the dynamic properties data.
+-   The "View Files" tab now no longer unnecessarily loads the dynamic properties data.
 
 # v1.0.0-beta.34